Making changes to AmbryMain and RestServerMain
Changes to AmbryMain and RestServerMain to make them consistent with each other so that
the experience of starting these components in open source is consistent.

Changes in RestServerMain:
1. log4j props argument removed because it can be set up as a VM option.
2. Return right exit code in case of exceptions.

Changes in AmbryMain:
1. Argument receiving.
2. Exit code handling.
3. Now consistent with RestServerMain
