+++ build.gradle
@@ -47,6 +47,7 @@ project(':ambry-utils') {
     dependencies {
         compile "commons-codec:commons-codec:$commonsVersion"
         compile "org.json:json:$jsonVersion"
+        compile "net.sf.jopt-simple:jopt-simple:$joptSimpleVersion"
     }
 }
