+++ config/frontend.properties
@@ -14,8 +14,7 @@
 rest.server.blob.storage.service.factory=com.github.ambry.frontend.AmbryBlobStorageServiceFactory
 
 # coordinator
+router.hostname=router
+router.datacenter.name=Data-Center
 coordinator.hostname=localhost
 coordinator.datacenter.name=Data-Center
-
-# netty
-netty.server.port=1174
